+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- Added MakeTablePath, CopyTable, CopyTables, DescribeTable methods for TableClient
+- Add logging for transactions
+
 ## v0.1.5
 - Fix timeout error on create session
 - Fix transport error on delete session